Adapt tests to strict- and warnings-by-default
In a future version of Perl, strictures and warnings may be on by
default.  These patches adapt the tests in this distribution to
anticipate that possibility.

Since Text-Balanced's Makefile.PL implies a minimum version of
perl-5.003_096, we don't include statements like 'use strict' or 'no
warnings' in these patches, as some aspects of the strict and warnings
pragmas did not exist as of that version of Perl.  We simply make the
corrections as needed.